Added Caching Features
